LambdaFunctions/Auth/requirements.txt
boto3==1.4.4
PyJWT==1.4.2